Improve sandworm_turnTowardTargetAnim (int turn-scale temp)
Type the >>2 turn-scale temp as int (not s16) and apply the (s16)
cast at the (f32) conversion, so the per-branch truncation extsh
collapse to the single join extsh the target emits. Drops two
spurious extsh. 17.6->29.2%; residual is one += store extsh plus
arg-eval order.

+++ src/main/dll/dll_014C_babycloudrunner.c
@@ -694,7 +694,7 @@ void sandworm_turnTowardTargetAnim(int a, int b, u8* c, int d)
         }
         else
         {
-            s16 t;
+            int t;
             s16 sv = (s16)shifted;
             if ((int)sv > 0)
             {
@@ -704,7 +704,7 @@ void sandworm_turnTowardTargetAnim(int a, int b, u8* c, int d)
             {
                 t = -sv >> 2;
             }
-            ((ObjAnimAdvanceObjectFirstF32Fn)ObjAnim_AdvanceCurrentMove)(a, (f32)t / lbl_803E4240, timeDelta, 0);
+            ((ObjAnimAdvanceObjectFirstF32Fn)ObjAnim_AdvanceCurrentMove)(a, (f32)(s16)t / lbl_803E4240, timeDelta, 0);
         }
     }
 }
